+++ ibCassandraModel/src/main/scala/model/datastax/ib/feed/request/RequestDataProvider.scala
@@ -33,7 +33,7 @@ class RequestDataProvider(
   def changeStateData(
     histData: RequestData,
     newState: RequestState,
-    rowsReceived: Option[Int] = None,
+    rowsReceived: Option[Long] = None,
     error: Option[String]     = None
   ): CompletionStage[Unit] = batch(
     Seq(
